+++ Patrullero/src/frsf/cidisi/exercise/patrullero/search/modelo/ObstaculoParcial.java
@@ -3,7 +3,7 @@ package frsf.cidisi.exercise.patrullero.search.modelo;
 public class ObstaculoParcial extends Obstaculo {
 	private Integer retardoMultiplicativo;
 
-	public ObstaculoParcial(Long id, NombreObstaculo nombre, Long tiempoInicio, Long tiempoFin, Visibilidad visibilidad, Lugar lugar, Integer retardoMultiplicativo) {
+	public ObstaculoParcial(Long id, NombreObstaculo nombre, Integer tiempoInicio, Integer tiempoFin, Visibilidad visibilidad, Lugar lugar, Integer retardoMultiplicativo) {
 		super(id, nombre, tiempoInicio, tiempoFin, visibilidad, lugar);
 		this.retardoMultiplicativo = retardoMultiplicativo;
 	}
