+++ Patrullero/src/frsf/cidisi/exercise/patrullero/search/modelo/ObstaculoTotal.java
@@ -2,7 +2,7 @@ package frsf.cidisi.exercise.patrullero.search.modelo;
 
 public class ObstaculoTotal extends Obstaculo {
 
-	public ObstaculoTotal(Long id, NombreObstaculo nombre, Long tiempoInicio, Long tiempoFin, Visibilidad visibilidad, Lugar lugar) {
+	public ObstaculoTotal(Long id, NombreObstaculo nombre, Integer tiempoInicio, Integer tiempoFin, Visibilidad visibilidad, Lugar lugar) {
 		super(id, nombre, tiempoInicio, tiempoFin, visibilidad, lugar);
 	}
